+++ template/AdminAuthoritySiret.php
@@ -5,7 +5,11 @@
 </p>
 <h2>Gestion des numéro SIRET</h2>
 
-<ul class="list-group col-xs-6">
+<div class="col-xs-6">
+<ul class="list-group">
+    <li class="list-group-item">
+        <b>SIRET activé</b>
+    </li>
 <?php if (! $siret_list): ?>
 	<li class="list-group-item">
     	Cette collectivité n'est associée à aucun numéro SIRET.
@@ -16,17 +20,69 @@
 	<li class="list-group-item">
 	  	<?php hecho($siret_info['siret'])?>
 	  	<?php if($this->me->isSuper()):?>
-	    <form action='admin_authority_siret_del.php' method='post' class="pull-right">    
-	    	<input type='hidden' name='authority_siret_id' value='<?php hecho($siret_info['id'])?>'/>
-			<button class="btn btn-xs btn-warning" type='submit'>
-			<span class="glyphicon glyphicon-trash"></span>
-			</button>
-		</form>  
+
+
+            <form action='admin_authority_siret_del.php' method='post' class="pull-right">
+                <input type='hidden' name='authority_siret_id' value='<?php hecho($siret_info['id'])?>'/>
+                <button class="btn btn-xs btn-warning" type='submit' title="Supprimer ce numéro SIRET">
+                <span class="glyphicon glyphicon-trash"></span>
+                </button>
+            </form>
+
+            <span class="pull-right">&nbsp;</span>
+
+            <form action='admin_authority_siret_block.php' method='post' class="pull-right">
+                <input type='hidden' name='authority_siret_id' value='<?php hecho($siret_info['id'])?>'/>
+                <button class="btn btn-xs btn-info" type='submit' title="Bloquer ce numéro SIRET">
+                    <span class="glyphicon glyphicon-thumbs-down"></span>
+                </button>
+            </form>
+
+
 		<?php endif;?>
 	</li>
 <?php endforeach;?>
+
+
+</ul>
+<ul class="list-group">
+<li class="list-group-item">
+    <b>SIRET bloqués</b>
+</li>
+
+	<?php if (! $siret_blocked_list): ?>
+        <li class="list-group-item">
+            Cette collectivité n'a pas de SIRET bloqué.
+        </li>
+	<?php endif;?>
+	<?php foreach($siret_blocked_list as $siret_info) : ?>
+        <li class="list-group-item">
+			<?php hecho($siret_info['siret'])?>
+			<?php if($this->me->isSuper()):?>
+
+                <form action='admin_authority_siret_del.php' method='post' class="pull-right">
+                    <input type='hidden' name='authority_siret_id' value='<?php hecho($siret_info['id'])?>'/>
+                    <button class="btn btn-xs btn-warning" type='submit' title="Supprimer ce numéro SIRET">
+                        <span class="glyphicon glyphicon-trash"></span>
+                    </button>
+                </form>
+
+                <span class="pull-right">&nbsp;</span>
+
+                <form action='admin_authority_siret_unblock.php' method='post' class="pull-right">
+                    <input type='hidden' name='authority_siret_id' value='<?php hecho($siret_info['id'])?>'/>
+                    <button class="btn btn-xs btn-info" type='submit' title="Débloquer ce numéro SIRET">
+                        <span class="glyphicon glyphicon-thumbs-up"></span>
+                    </button>
+                </form>
+
+
+			<?php endif;?>
+        </li>
+	<?php endforeach;?>
 </ul>
 
+</div>
 
 <div class='col-xs-6'>
 <?php if($this->me->isSuper()):?>
@@ -43,7 +99,9 @@
 <?php endif;?>
 <div class="panel panel-default">
   <div class="panel-body">
-	Ces numéros SIRET sont utilisés dans le cadre du protocole PES pour associer les PES Retour en provenance d'Hélios à la collectivité.
+      <p>Ces numéros SIRET sont utilisés dans le cadre du protocole PES pour associer les PES Retour en provenance d'Hélios à la collectivité.</p>
+      <p>Lorsqu'un PES Retour arrive on selectionne la collectivité avec le <b>SIRET activé</b></p>
+      <p>Les <b>SIRET bloqué</b> permettent de ne pas selectionner la collectivité (erreur dans le numéro SIRET sur le PES Aller)</p>
   </div>
 </div>