+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Export de la liste des collectvités pour le super admin ou l'admin de groupe #300
 - Possibilité de tester les certificats (PEM) pour valider s'ils sont RGS et/ou utilisé sur la plateforme #366
 - Ajout d'un bouton permettant de poster l'acte sans le signer s'il est en attente de signature #374
+- Possibilité de bloquer des numéro SIRET afin d'éviter les erreurs d'acheminement des PES Retour #372
 
 ### Retraits